arrow_row::fixed

Trait FixedLengthEncoding

pub trait FixedLengthEncoding: Copy {
    type Encoded: Sized + Copy + FromSlice + AsRef<[u8]> + AsMut<[u8]>;

    const ENCODED_LEN: usize = _;

    // Required methods
    fn encode(self) -> Self::Encoded;
    fn decode(encoded: Self::Encoded) -> Self;
}

Encodes a value of a particular fixed width type into bytes according to the rules described on super::RowConverter

Provided Associated Constants

const ENCODED_LEN: usize = _

Required Associated Types

type Encoded: Sized + Copy + FromSlice + AsRef<[u8]> + AsMut<[u8]>

Required Methods

fn encode(self) -> Self::Encoded

fn decode(encoded: Self::Encoded) -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.