Struct arrow_row::Rows

pub struct Rows { /* private fields */ }

A row-oriented representation of arrow data, that is normalized for comparison.

See the module level documentation and RowConverter for more details.

Implementations

impl Rows

pub fn push(&mut self, row: Row<'_>)

Append a Row to this Rows

pub fn row(&self, row: usize) -> Row<'_>

Returns the row at index row

pub fn clear(&mut self)

Sets the length of this Rows to 0

pub fn num_rows(&self) -> usize

Returns the number of Row in this Rows

pub fn iter(&self) -> RowsIter<'_>

Returns an iterator over the Row in this Rows

pub fn size(&self) -> usize

Returns the size of this instance in bytes

Includes the size of Self.