Module arrow_row::variable

Constants

BLOCK_CONTINUATION

The continuation token

BLOCK_SIZE

The block size of the variable length encoding

EMPTY_SENTINEL

Indicates an empty string

NON_EMPTY_SENTINEL

Indicates a non-empty string

Functions

decode_binary

Decodes a binary array from rows with the provided options

decode_string⚠

Decodes a string array from rows with the provided options

decoded_len 🔒

Returns the number of bytes of encoded data

encode

Variable length values are encoded as

encode_one

encoded_len

Returns the length of the encoded representation of a byte array, including the null byte

padded_length

Returns the padded length of the encoded length of the given length