Module arrow_row::fixed

Macros

encode_signed 🔒

encode_unsigned 🔒

Traits

FixedLengthEncoding

Encodes a value of a particular fixed width type into bytes according to the rules described on super::RowConverter

FromSlice

Functions

decode_bool

Decodes a BooleanArray from rows

decode_fixed 🔒 ⚠

Decodes a ArrayData from rows based on the provided FixedLengthEncoding T

decode_fixed_size_binary

Decodes a FixedLengthBinary from rows

decode_nulls

Decodes a single byte from each row, interpreting 0x01 as a valid value and all other values as a null

decode_primitive

Decodes a PrimitiveArray from rows

encode

Fixed width types are encoded as

encode_boolean

Boolean values are encoded as

encode_boolean_not_null

Encoding for non-nullable boolean arrays. Iterates directly over values, and skips NULLs-checking.

encode_fixed_size_binary

encode_not_null

Encoding for non-nullable primitive arrays. Iterates directly over the values, and skips NULLs-checking.

encoded_len

Returns the total encoded length (including null byte) for a value of type T::Native

split_off 🔒

Splits len bytes from src