Module arrow_row::variable

Constants

BLOCK_CONTINUATION

The continuation token

BLOCK_SIZE

The block size of the variable length encoding

EMPTY_SENTINEL

Indicates an empty string

MINI_BLOCK_COUNT

The first block is split into MINI_BLOCK_COUNT mini-blocks

MINI_BLOCK_SIZE

The mini block size

NON_EMPTY_SENTINEL

Indicates a non-empty string

Functions

decode_binary

Decodes a binary array from rows with the provided options

decode_binary_view

Decodes a binary view array from rows with the provided options

decode_blocks

decode_string⚠

Decodes a string array from rows with the provided options

decode_string_view⚠

Decodes a string view array from rows with the provided options

decoded_len 🔒

Returns the number of bytes of encoded data

encode

Variable length values are encoded as

encode_blocks 🔒

Writes val in SIZE blocks with the appropriate continuation tokens

encode_empty

encode_null

encode_one

encoded_len

Returns the length of the encoded representation of a byte array, including the null byte

padded_length

Returns the padded length of the encoded length of the given length