arrow_row

Module variable

Constants

BLOCK_CONTINUATION

The continuation token

BLOCK_SIZE

The block size of the variable length encoding

EMPTY_SENTINEL

Indicates an empty string

MINI_BLOCK_COUNT

The first block is split into MINI_BLOCK_COUNT mini-blocks

MINI_BLOCK_SIZE

The mini block size

NON_EMPTY_SENTINEL

Indicates a non-empty string

NULL_VALUE_SENTINEL

Indicates a Null value (for DataType::Null)

Functions

decode_binary

Decodes a binary array from rows with the provided options

decode_binary_view

Decodes a binary view array from rows with the provided options

decode_binary_view_inner 🔒

decode_blocks

Decodes a single block of data The f function accepts a slice of the decoded data, it may be called multiple times

decode_null_value

decode_string⚠

Decodes a string array from rows with the provided options

decode_string_view⚠

Decodes a string view array from rows with the provided options

decoded_len 🔒

Returns the number of bytes of encoded data

encode

Variable length values are encoded as

encode_blocks 🔒

Writes val in SIZE blocks with the appropriate continuation tokens

encode_empty

encode_generic_byte_array 🔒

Calls encode with optimized iterator for generic byte arrays

encode_null

encode_null_value

Ensure NullArrays don’t get encoded as empty lists which can lose their length

encode_one

non_null_padded_length 🔒

Returns the padded length of the encoded length of the given length

padded_length

Returns the padded length of the encoded length of the given length